refactor: rename module to material_importer_service
